#85668 Fix an issue with SB token expiration (#265)
* #85668 Fix an issue with SB token expiration

* Remove unneeded library

* Fix build
