+++ src/CaptainHook.Common/CaptainHook.Common.csproj
@@ -20,7 +20,6 @@
     <PackageReference Include="Microsoft.Azure.Management.Fluent" Version="1.34.0" />
   	<PackageReference Include="Microsoft.Azure.ServiceBus" Version="4.1.4" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="3.1.7" />
-    <PackageReference Include="Nito.AsyncEx.Coordination" Version="5.0.0" />
     <PackageReference Include="Polly" Version="7.2.1" />
   </ItemGroup>
