+++ src/CaptainHook.DirectorService/Program.cs
@@ -18,16 +18,7 @@ using CaptainHook.DirectorService.ReaderServiceManagement;
 using CaptainHook.Storage.Cosmos;
 using Eshopworld.Data.CosmosDb.Extensions;
 using Eshopworld.Telemetry;
-using Microsoft.Azure.Management.ResourceManager.Fluent;
-using Microsoft.Azure.Management.ResourceManager.Fluent.Authentication;
-using Microsoft.Azure.Management.ResourceManager.Fluent.Core;
-using Microsoft.Azure.Management.ServiceBus.Fluent;
-using Microsoft.Azure.Services.AppAuthentication;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Rest;
-using Nito.AsyncEx;
-using IServiceBusManager = CaptainHook.Common.ServiceBus.IServiceBusManager;
-using ServiceBusManager = CaptainHook.Common.ServiceBus.ServiceBusManager;
 
 namespace CaptainHook.DirectorService
 {
@@ -104,8 +95,6 @@ namespace CaptainHook.DirectorService
                 builder.RegisterType<MessageProviderFactory>()
                     .As<IMessageProviderFactory>()
                     .SingleInstance();
-
-                builder.ConfigureServiceBusManager(configurationSettings);
 
                 builder.RegisterType<ServiceBusManager>()
                     .As<IServiceBusManager>();
